+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Bugfix: `crate()` is using the correct 'topenv' environment now.
 * Remove the unused 'safe' variants of dictionary getters
+* `dictionary_sugar_get()` and corresponding functions now take a list of dictionaries as 
+  optional argument `.dicts_suggest` to look for suggestions if `.key` is not part of the dictionary.
 
 # mlr3misc 0.15.1